python.streamlit/requirements.txt
streamlit
pandas
babel
oauth2client
gspread
numpy
matplotlib
plotly
python-dotenv
